+++ cmd/dm/command/deploy.go
@@ -15,16 +15,15 @@ package command
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 
 	"github.com/fatih/color"
 	"github.com/joomcode/errorx"
 	"github.com/pingcap-incubator/tiup-cluster/pkg/bindversion"
 	"github.com/pingcap-incubator/tiup-cluster/pkg/cliutil"
+	"github.com/pingcap-incubator/tiup-cluster/pkg/cliutil/prepare"
 	"github.com/pingcap-incubator/tiup-cluster/pkg/clusterutil"
 	"github.com/pingcap-incubator/tiup-cluster/pkg/errutil"
 	"github.com/pingcap-incubator/tiup-cluster/pkg/log"
@@ -32,23 +31,20 @@ import (
 	"github.com/pingcap-incubator/tiup-cluster/pkg/meta"
 	"github.com/pingcap-incubator/tiup-cluster/pkg/task"
 	"github.com/pingcap-incubator/tiup-cluster/pkg/utils"
+	"github.com/pingcap-incubator/tiup/pkg/set"
 	tiuputils "github.com/pingcap-incubator/tiup/pkg/utils"
 	"github.com/pingcap/errors"
 	"github.com/spf13/cobra"
-	"go.uber.org/zap"
 )
 
 var (
 	errNSDeploy            = errNS.NewSubNamespace("deploy")
 	errDeployNameDuplicate = errNSDeploy.NewType("name_dup", errutil.ErrTraitPreCheck)
-	errDeployDirConflict   = errNSDeploy.NewType("dir_conflict", errutil.ErrTraitPreCheck)
-	errDeployPortConflict  = errNSDeploy.NewType("port_conflict", errutil.ErrTraitPreCheck)
 )
 
 type deployOptions struct {
 	user         string // username to login to the SSH server
 	identityFile string // path to the private key file
-	skipConfirm  bool   // skip the confirmation of topology
 }
 
 func newDeploy() *cobra.Command {
@@ -74,238 +70,11 @@ func newDeploy() *cobra.Command {
 
 	cmd.Flags().StringVar(&opt.user, "user", "root", "The user name to login via SSH. The user must has root (or sudo) privilege.")
 	cmd.Flags().StringVarP(&opt.identityFile, "identity_file", "i", "", "The path of the SSH identity file. If specified, public key authentication will be used.")
-	cmd.Flags().BoolVarP(&opt.skipConfirm, "yes", "y", false, "Skip confirming the topology")
 
 	return cmd
 }
 
-func fixDir(topo *meta.DMSpecification) func(string) string {
-	return func(dir string) string {
-		if dir != "" {
-			return clusterutil.Abs(topo.GlobalOptions.User, dir)
-		}
-		return dir
-	}
-}
-
-func checkClusterDirConflict(clusterName string, topo *meta.DMSpecification) error {
-	type DirAccessor struct {
-		dirKind  string
-		accessor func(meta.Instance, *meta.DMTopologySpecification) string
-	}
-
-	instanceDirAccessor := []DirAccessor{
-		{dirKind: "deploy directory", accessor: func(instance meta.Instance, topo *meta.DMTopologySpecification) string { return instance.DeployDir() }},
-		{dirKind: "data directory", accessor: func(instance meta.Instance, topo *meta.DMTopologySpecification) string { return instance.DataDir() }},
-		{dirKind: "log directory", accessor: func(instance meta.Instance, topo *meta.DMTopologySpecification) string { return instance.LogDir() }},
-	}
-	hostDirAccessor := []DirAccessor{
-		{dirKind: "monitor deploy directory", accessor: func(instance meta.Instance, topo *meta.DMTopologySpecification) string {
-			return topo.MonitoredOptions.DeployDir
-		}},
-		{dirKind: "monitor data directory", accessor: func(instance meta.Instance, topo *meta.DMTopologySpecification) string {
-			return topo.MonitoredOptions.DataDir
-		}},
-		{dirKind: "monitor log directory", accessor: func(instance meta.Instance, topo *meta.DMTopologySpecification) string {
-			return topo.MonitoredOptions.LogDir
-		}},
-	}
-
-	type Entry struct {
-		clusterName string
-		dirKind     string
-		dir         string
-		instance    meta.Instance
-	}
-
-	currentEntries := []Entry{}
-	existingEntries := []Entry{}
-
-	clusterDir := meta.ProfilePath(meta.TiOpsClusterDir)
-	fileInfos, err := ioutil.ReadDir(clusterDir)
-	if err != nil && !os.IsNotExist(err) {
-		return err
-	}
-	for _, fi := range fileInfos {
-		if fi.Name() == clusterName {
-			continue
-		}
-
-		if tiuputils.IsNotExist(meta.ClusterPath(fi.Name(), meta.MetaFileName)) {
-			continue
-		}
-		metadata, err := meta.DMMetadata(fi.Name())
-		if err != nil {
-			return errors.Trace(err)
-		}
-
-		f := fixDir(metadata.Topology)
-		metadata.Topology.IterInstance(func(inst meta.Instance) {
-			for _, dirAccessor := range instanceDirAccessor {
-				existingEntries = append(existingEntries, Entry{
-					clusterName: fi.Name(),
-					dirKind:     dirAccessor.dirKind,
-					dir:         f(dirAccessor.accessor(inst, metadata.Topology)),
-					instance:    inst,
-				})
-			}
-		})
-		metadata.Topology.IterHost(func(inst meta.Instance) {
-			for _, dirAccessor := range hostDirAccessor {
-				existingEntries = append(existingEntries, Entry{
-					clusterName: fi.Name(),
-					dirKind:     dirAccessor.dirKind,
-					dir:         f(dirAccessor.accessor(inst, metadata.Topology)),
-					instance:    inst,
-				})
-			}
-		})
-	}
-
-	f := fixDir(topo)
-	topo.IterInstance(func(inst meta.Instance) {
-		for _, dirAccessor := range instanceDirAccessor {
-			currentEntries = append(currentEntries, Entry{
-				dirKind:  dirAccessor.dirKind,
-				dir:      f(dirAccessor.accessor(inst, topo)),
-				instance: inst,
-			})
-		}
-	})
-	topo.IterHost(func(inst meta.Instance) {
-		for _, dirAccessor := range hostDirAccessor {
-			currentEntries = append(currentEntries, Entry{
-				dirKind:  dirAccessor.dirKind,
-				dir:      f(dirAccessor.accessor(inst, topo)),
-				instance: inst,
-			})
-		}
-	})
-
-	for _, d1 := range currentEntries {
-		for _, d2 := range existingEntries {
-			if d1.instance.GetHost() != d2.instance.GetHost() {
-				continue
-			}
-
-			if d1.dir == d2.dir && d1.dir != "" {
-				properties := map[string]string{
-					"ThisDirKind":    d1.dirKind,
-					"ThisDir":        d1.dir,
-					"ThisComponent":  d1.instance.ComponentName(),
-					"ThisHost":       d1.instance.GetHost(),
-					"ExistCluster":   d2.clusterName,
-					"ExistDirKind":   d2.dirKind,
-					"ExistDir":       d2.dir,
-					"ExistComponent": d2.instance.ComponentName(),
-					"ExistHost":      d2.instance.GetHost(),
-				}
-				zap.L().Info("Meet deploy directory conflict", zap.Any("info", properties))
-				return errDeployDirConflict.New("Deploy directory conflicts to an existing cluster").WithProperty(cliutil.SuggestionFromTemplate(`
-The directory you specified in the topology file is:
-  Directory: {{ColorKeyword}}{{.ThisDirKind}} {{.ThisDir}}{{ColorReset}}
-  Component: {{ColorKeyword}}{{.ThisComponent}} {{.ThisHost}}{{ColorReset}}
-
-It conflicts to a directory in the existing cluster:
-  Existing Cluster Name: {{ColorKeyword}}{{.ExistCluster}}{{ColorReset}}
-  Existing Directory:    {{ColorKeyword}}{{.ExistDirKind}} {{.ExistDir}}{{ColorReset}}
-  Existing Component:    {{ColorKeyword}}{{.ExistComponent}} {{.ExistHost}}{{ColorReset}}
-
-Please change to use another directory or another host.
-`, properties))
-			}
-		}
-	}
-
-	return nil
-}
-
-func checkClusterPortConflict(clusterName string, topo *meta.DMTopologySpecification) error {
-	clusterDir := meta.ProfilePath(meta.TiOpsClusterDir)
-	fileInfos, err := ioutil.ReadDir(clusterDir)
-	if err != nil && !os.IsNotExist(err) {
-		return err
-	}
-
-	type Entry struct {
-		clusterName string
-		instance    meta.Instance
-		port        int
-	}
-
-	currentEntries := []Entry{}
-	existingEntries := []Entry{}
-
-	for _, fi := range fileInfos {
-		if fi.Name() == clusterName {
-			continue
-		}
-
-		if tiuputils.IsNotExist(meta.ClusterPath(fi.Name(), meta.MetaFileName)) {
-			continue
-		}
-		metadata, err := meta.ClusterMetadata(fi.Name())
-		if err != nil {
-			return errors.Trace(err)
-		}
-
-		metadata.Topology.IterInstance(func(inst meta.Instance) {
-			for _, port := range inst.UsedPorts() {
-				existingEntries = append(existingEntries, Entry{
-					clusterName: fi.Name(),
-					instance:    inst,
-					port:        port,
-				})
-			}
-		})
-	}
-
-	topo.IterInstance(func(inst meta.Instance) {
-		for _, port := range inst.UsedPorts() {
-			currentEntries = append(currentEntries, Entry{
-				instance: inst,
-				port:     port,
-			})
-		}
-	})
-
-	for _, p1 := range currentEntries {
-		for _, p2 := range existingEntries {
-			if p1.instance.GetHost() != p2.instance.GetHost() {
-				continue
-			}
-
-			if p1.port == p2.port {
-				properties := map[string]string{
-					"ThisPort":       strconv.Itoa(p1.port),
-					"ThisComponent":  p1.instance.ComponentName(),
-					"ThisHost":       p1.instance.GetHost(),
-					"ExistCluster":   p2.clusterName,
-					"ExistPort":      strconv.Itoa(p2.port),
-					"ExistComponent": p2.instance.ComponentName(),
-					"ExistHost":      p2.instance.GetHost(),
-				}
-				zap.L().Info("Meet deploy port conflict", zap.Any("info", properties))
-				return errDeployPortConflict.New("Deploy port conflicts to an existing cluster").WithProperty(cliutil.SuggestionFromTemplate(`
-The port you specified in the topology file is:
-  Port:      {{ColorKeyword}}{{.ThisPort}}{{ColorReset}}
-  Component: {{ColorKeyword}}{{.ThisComponent}} {{.ThisHost}}{{ColorReset}}
-
-It conflicts to a port in the existing cluster:
-  Existing Cluster Name: {{ColorKeyword}}{{.ExistCluster}}{{ColorReset}}
-  Existing Port:         {{ColorKeyword}}{{.ExistPort}}{{ColorReset}}
-  Existing Component:    {{ColorKeyword}}{{.ExistComponent}} {{.ExistHost}}{{ColorReset}}
-
-Please change to use another port or another host.
-`, properties))
-			}
-		}
-	}
-
-	return nil
-}
-
-func confirmTopology(clusterName, version string, topo *meta.DMTopologySpecification) error {
+func confirmTopology(clusterName, version string, topo *meta.DMTopologySpecification, patchedRoles set.StringSet) error {
 	log.Infof("Please confirm your topology:")
 
 	cyan := color.New(color.FgCyan, color.Bold)
@@ -318,8 +87,12 @@ func confirmTopology(clusterName, version string, topo *meta.DMTopologySpecifica
 	}
 
 	topo.IterInstance(func(instance meta.Instance) {
+		comp := instance.ComponentName()
+		if patchedRoles.Exist(comp) {
+			comp = comp + " (patched)"
+		}
 		clusterTable = append(clusterTable, []string{
-			instance.ComponentName(),
+			comp,
 			instance.GetHost(),
 			utils.JoinInt(instance.UsedPorts(), "/"),
 			strings.Join(instance.UsedDirs(), ","),
@@ -330,7 +103,10 @@ func confirmTopology(clusterName, version string, topo *meta.DMTopologySpecifica
 
 	log.Warnf("Attention:")
 	log.Warnf("    1. If the topology is not what you expected, check your yaml file.")
-	log.Warnf("    1. Please confirm there is no port/directory conflicts in same host.")
+	log.Warnf("    2. Please confirm there is no port/directory conflicts in same host.")
+	if len(patchedRoles) != 0 {
+		log.Errorf("    3. The component marked as `patched` has been replaced by previours patch command.")
+	}
 
 	return cliutil.PromptForConfirmOrAbortError("Do you want to continue? [y/N]: ")
 }
@@ -351,15 +127,15 @@ func deploy(clusterName, clusterVersion, topoFile string, opt deployOptions) err
 		return err
 	}
 
-	if err := checkClusterPortConflict(clusterName, &topo); err != nil {
+	if err := prepare.CheckClusterPortConflict(clusterName, &topo); err != nil {
 		return err
 	}
-	if err := checkClusterDirConflict(clusterName, &topo); err != nil {
+	if err := prepare.CheckClusterDirConflict(clusterName, &topo); err != nil {
 		return err
 	}
 
-	if !opt.skipConfirm {
-		if err := confirmTopology(clusterName, clusterVersion, &topo); err != nil {
+	if !skipConfirm {
+		if err := confirmTopology(clusterName, clusterVersion, &topo, set.NewStringSet()); err != nil {
 			return err
 		}
 	}
@@ -414,7 +190,7 @@ func deploy(clusterName, clusterVersion, topoFile string, opt deployOptions) err
 	})
 
 	// Download missing component
-	downloadCompTasks = buildDownloadCompTasks(clusterVersion, &topo)
+	downloadCompTasks = prepare.BuildDownloadCompTasks(clusterVersion, &topo)
 
 	// Deploy components to remote
 	topo.IterInstance(func(inst meta.Instance) {
@@ -479,20 +255,4 @@ func deploy(clusterName, clusterVersion, topoFile string, opt deployOptions) err
 	hint := color.New(color.Bold).Sprintf("tiup dm start %s", clusterName)
 	log.Infof("Deployed cluster `%s` successfully, you can start the cluster via `%s`", clusterName, hint)
 	return nil
-}
-
-func buildDownloadCompTasks(version string, topo *meta.DMTopologySpecification) []*task.StepDisplay {
-	var tasks []*task.StepDisplay
-	topo.IterComponent(func(comp meta.Component) {
-		if len(comp.Instances()) < 1 {
-			return
-		}
-		version := bindversion.ComponentVersion(comp.Name(), version)
-		t := task.
-			NewBuilder().
-			Download(comp.Name(), version).
-			BuildAsStep(fmt.Sprintf("  - Download %s:%s", comp.Name(), version))
-		tasks = append(tasks, t)
-	})
-	return tasks
 }
