+++ cmd/dm/command/root.go
@@ -16,7 +16,6 @@ package command
 import (
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 
 	"github.com/fatih/color"
@@ -30,7 +29,6 @@ import (
 	"github.com/pingcap-incubator/tiup-cluster/pkg/version"
 	"github.com/pingcap-incubator/tiup/pkg/localdata"
 	tiupmeta "github.com/pingcap-incubator/tiup/pkg/meta"
-
 	"github.com/pingcap-incubator/tiup/pkg/repository"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
@@ -39,8 +37,9 @@ import (
 var rootCmd *cobra.Command
 
 var (
-	errNS      = errorx.NewNamespace("cmd")
-	sshTimeout int64 // timeout in seconds when connecting an SSH server
+	errNS       = errorx.NewNamespace("cmd")
+	sshTimeout  int64 // timeout in seconds when connecting an SSH server
+	skipConfirm bool
 )
 
 func init() {
@@ -53,7 +52,7 @@ func init() {
 	cobra.EnableCommandSorting = false
 
 	rootCmd = &cobra.Command{
-		Use:           filepath.Base(os.Args[0]),
+		Use:           cliutil.OsArgs0(),
 		Short:         "Deploy a DM cluster for production",
 		SilenceUsage:  true,
 		SilenceErrors: true,
@@ -75,9 +74,10 @@ func init() {
 		},
 	}
 
-	beautifyCobraUsageAndHelp(rootCmd)
+	cliutil.BeautifyCobraUsageAndHelp(rootCmd)
 
 	rootCmd.PersistentFlags().Int64Var(&sshTimeout, "ssh-timeout", 5, "Timeout in seconds to connect host via SSH, ignored for operations that don't need an SSH connection.")
+	rootCmd.PersistentFlags().BoolVarP(&skipConfirm, "yes", "y", false, "Skip all confirmations and assumes 'yes'")
 
 	rootCmd.AddCommand(
 		newDeploy(),
@@ -187,31 +187,4 @@ func Execute() {
 	if code != 0 {
 		os.Exit(code)
 	}
-}
-
-func beautifyCobraUsageAndHelp(rootCmd *cobra.Command) {
-	rootCmd.SetUsageTemplate(`Usage:{{if .Runnable}}
-  {{ColorCommand}}{{.UseLine}}{{ColorReset}}{{end}}{{if .HasAvailableSubCommands}}
-  {{ColorCommand}}{{.CommandPath}} [command]{{ColorReset}}{{end}}{{if gt (len .Aliases) 0}}
-
-Aliases:
-  {{ColorCommand}}{{.NameAndAliases}}{{ColorReset}}{{end}}{{if .HasExample}}
-
-Examples:
-{{.Example}}{{end}}{{if .HasAvailableSubCommands}}
-
-Available Commands:{{range .Commands}}{{if (or .IsAvailableCommand (eq .Name "help"))}}
-  {{rpad .Name .NamePadding }} {{.Short}}{{end}}{{end}}{{end}}{{if .HasAvailableLocalFlags}}
-
-Flags:
-{{.LocalFlags.FlagUsages | trimTrailingWhitespaces}}{{end}}{{if .HasAvailableInheritedFlags}}
-
-Global Flags:
-{{.InheritedFlags.FlagUsages | trimTrailingWhitespaces}}{{end}}{{if .HasHelpSubCommands}}
-
-Additional help topics:{{range .Commands}}{{if .IsAdditionalHelpTopicCommand}}
-  {{rpad .CommandPath .CommandPathPadding}} {{.Short}}{{end}}{{end}}{{end}}{{if .HasAvailableSubCommands}}
-
-Use "{{ColorCommand}}{{.CommandPath}} [command] --help{{ColorReset}}" for more information about a command.{{end}}
-`)
 }
